Bug 32409: change Unsigned16s in RelayBitmap_t to Unsigned8

Related work items: #32409

+++ device/src/main/java/a75f/io/device/mesh/LSmartNode.java
@@ -272,7 +272,7 @@ public class LSmartNode
             airflowCFMIntegralKFactor = (int) (100 * acbEquip.getVavAirflowCFMIntegralKFactor().readPriorityVal()); // fallback
             enableCFM = (int) (acbEquip.getEnableCFMControl().readPriorityVal());
 
-            settings2.relayBitmap.relay1.set((int) acbEquip.getRelay1OutputAssociation().readDefaultVal());
+            settings2.relayBitmap.relay1.set((short)acbEquip.getRelay1OutputAssociation().readDefaultVal());
             // Relay 2 will always be zero, because relay 2 is not used in ACB and no relay 2 association point will exist
 
         } else if (isEquipType("vav", address)) {
